Seed db finish

+++ server/db/models/lineItem.js
@@ -8,6 +8,6 @@ module.exports = db.define('lineItem', {
     type: Sequelize.INTEGER
   },
   price: {
-    type: Sequelize.DECIMAL
+    type: Sequelize.INTEGER
   }
 });
